apps/available-schedules-python/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
prometheus_client==0.21.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0